move file to utils folder
